requirements.txt
pwntools
requests
pycryptodome
python-magic
colorama
